refactor: Merge Box's pharaoh child class into the parent one
